test: enable IndicatorView direct-jump test on all platforms
Removes the iOS-only `#if TEST_FAILS_ON_ANDROID && TEST_FAILS_ON_WINDOWS
&& TEST_FAILS_ON_CATALYST` guard and the `[Ignore]` attribute so the
test now runs on every platform test project. Android and Windows
exercise the direct-jump behavior and pass; iOS and Mac Catalyst fail
with the documented `Pos:1` observation from #27007, turning a silent
`[Ignore]` into a live red signal that will flip green once the
UIPageControl-backed IndicatorView maps taps to the tapped dot index.

Renames the symbols to drop the "iOS" qualifier and widens
`PlatformAffected` to `iOS | macOS`. Waits for the CarouselView's
AutomationId instead of item text so the wait resolves on Android
where Label text isn't queryable without its own AutomationId.
