--- src/Controls/tests/TestCases.Shared.Tests/Tests/Issues/CarouselViewUITests.IndicatorViewiOSTapDirectJump.cs
@@ -1,50 +1,0 @@
-#if TEST_FAILS_ON_ANDROID && TEST_FAILS_ON_WINDOWS && TEST_FAILS_ON_CATALYST // iOS-only: the bug is specific to UIKit UIPageControl.
-using NUnit.Framework;
-using UITest.Appium;
-using UITest.Core;
-
-namespace Microsoft.Maui.TestCases.Tests.Issues
-{
-	public class IndicatorViewiOSTapDirectJump : _IssuesUITest
-	{
-		const string IndicatorId = "jumpIndicator";
-		const string PositionLabelId = "jumpPositionLabel";
-
-		public IndicatorViewiOSTapDirectJump(TestDevice device)
-			: base(device) { }
-
-		public override string Issue =>
-			"iOS IndicatorView dot tap only advances by plus/minus 1 instead of jumping directly to the tapped dot";
-
-		// Repros https://github.com/dotnet/maui/issues/27007.
-		[Test]
-		[Category(UITestCategories.IndicatorView)]
-		[Ignore(
-			"Fails on iOS pending fix for https://github.com/dotnet/maui/issues/27007. On origin/main this assertion observes Pos:1 instead of the expected Pos:4 because UIKit's UIPageControl (backing IndicatorView) only reports left-half vs right-half taps, so the carousel only advances by 1 regardless of which dot was tapped."
-		)]
-		public void TappingLastDotJumpsDirectlyToLastItem()
-		{
-			App.WaitForElement("Item 0", timeout: TimeSpan.FromSeconds(30));
-			App.WaitForElement(IndicatorId);
-
-			// The carousel has 5 items. Tapping the rightmost dot should jump
-			// directly to Item 4. On iOS, UIPageControl only interprets taps as
-			// "left half = back 1" / "right half = forward 1", so the carousel
-			// only advances to Item 1.
-			var indicatorRect = App.FindElement(IndicatorId).GetRect();
-			var lastDotX = indicatorRect.X + indicatorRect.Width - 10;
-			var centerY = indicatorRect.Y + indicatorRect.Height / 2;
-			App.TapCoordinates(lastDotX, centerY);
-
-			Thread.Sleep(800);
-
-			var posText = App.FindElement(PositionLabelId).GetText();
-			Assert.That(
-				posText,
-				Is.EqualTo("Pos:4"),
-				"Tapping the last indicator dot must jump the CarouselView directly to the last item (Pos:4). On iOS, the current UIPageControl-based IndicatorView only advances by 1 (Pos:1), which is the bug under test."
-			);
-		}
-	}
-}
-#endif
